diffie hellman first try
- Generating private/public keys pair for client and -server;
- Client computes the shared key and sends it to the server;
- Server computes the shared key and sends it back to the client.

PROBLEM
The client and the server shared keys should be the same, but they are not! 

I think the problem is the CryptoFunctions.parameters attribute, that should the same on this class usage at the client and the server, but it is not! Sent message to teacher asking if he could understand.