generating parameters on a separate script to file loaded by server
As the parameters generation is a time expensive operation and in order to speed up the development process, it was removed from server initialization to a separate script, that only needs to be ran once.

When started, the server loads the parameters from the file ./parameters and creates his keys pair based on that.

On production the previous method (new parameters every time the server starts) must be recovered!